fix valor

+++ database/migrations/2020_07_04_001812_create_finances_table.php
@@ -19,7 +19,7 @@ class CreateFinancesTable extends Migration
             $table->integer('codCliente');
             $table->integer('fatura')->unique();
             $table->string('descricao');
-            $table->double('valor',8,2);
+            $table->decimal('valor',10,2);
             $table->string('cdreceita');
             $table->integer('operacao');
             $table->date('datacompetencia');
